Changed: Moved some methods related to adaptive simulation from SIMoutput to SIMgeneric such that SIMoutput only deals with result output, and not much else. Retained virtual interfaces for some methods that are used by AdaptiveSIM.
